feat(Auth): 회원 가입 버그 수정 및 인증 관련 응답 형식 개선
- 회원 가입 시 entity 저장이 안 되고 있는 버그 수정
- CreatedDate, LastModifiedDate 어노테이션 컬럼 자동 입력을 위해 EntityListeners 와 EnableJpaAuditing 어노테이션 등록 필요
- 로그인 시 쿠키에 토큰 값 넣어주도록 수정
- 회원가입, 로그인 API의 응답 형식 DTO로 규격화

+++ src/main/java/com/nimble/server_spring/infra/jwt/AuthTokenProvider.java
@@ -15,14 +15,14 @@ import java.util.List;
 public class AuthTokenProvider {
     private final Key accessTokenKey;
     private final Key refreshTokenKey;
-    private final Long accessTokenExpiry;
-    private final Long refreshTokenExpiry;
+    private final Integer accessTokenExpiry;
+    private final Integer refreshTokenExpiry;
 
     public AuthTokenProvider(
             String accessTokenSecret,
             String refreshTokenSecret,
-            Long accessTokenExpiry,
-            Long refreshTokenExpiry
+            Integer accessTokenExpiry,
+            Integer refreshTokenExpiry
     ) {
         this.accessTokenKey = Keys.hmacShaKeyFor(accessTokenSecret.getBytes());
         this.refreshTokenKey = Keys.hmacShaKeyFor(refreshTokenSecret.getBytes());
